Fix weird infinite scrolling bugs
Fixes some bugs with infinite scrolling by taking away the "all images
on page have loaded" requirement that was introduced to combat collapsed
divs (which we believe is fixed now due to other CSS fixes). This logic
wasn't sound because when filtering, it's possible not all images would
reload, thus not firing the handler appropriately to update the number
of images loaded (and sometimes causing Infinite scroll to never fire).