Correct List Filtering Part 1
Corrects the implementation of smart lists so that the following
conditions hold:

1. All parameters that the user can sort on from Explore page can be
applied on a smart lisst
2. Filters that override list rules are applied correctly

This is done by reimplementing smart list building to use the same
internal API we use for all searches.

Additionally, this fixes some frontend issues:

1. Default filters for smart lists are set to the lists rules. This
fixes the "reset" functionality and also ensures that AllFilters
displays the correct initial state on list load
2. Remove netflix-kids and hbo-go implicit network filtering options
3. Ensure that cast filtering value is correctly pre-popularted in the
autocomplete box on smart lists.

This change is Part 1, because non-smart lists still exhibit several of
the issues described here, however the implementation uses a different
ES inded and will be separated out.